refactor(IA): modify information architecture
