--- docs/_templates/fullpage.njk
@@ -1,7 +1,0 @@
-{% extends 'application.njk' %}
-
-{% block layout %}
-  {% block content %}
-    {# page content goes here #}
-  {% endblock %}
-{% endblock %}
