+++ docs/_templates/guide.njk
