--- docs/components/breadcrumb/index.html
@@ -1,49 +1,0 @@
----
-title: Breadcrumb
-assets:
-  - helix-ui.css
----
-{% extends 'component.njk' %}
-{% block content %}
-<!-- explanation goes here -->
-
-<div class="demo">
-  <nav class="hxBreadcrumb">
-    <a href="#">Home</a>
-  </nav>
-
-  <nav class="hxBreadcrumb">
-    <a href="#">Home</a>
-    <hx-icon class="delimiter" type="angle-right"></hx-icon>
-    <a href="#">Library</a>
-  </nav>
-
-  <nav class="hxBreadcrumb">
-    <a href="#">Home</a>
-    <hx-icon class="delimiter" type="angle-right"></hx-icon>
-    <a href="#">Library</a>
-    <hx-icon class="delimiter" type="angle-right"></hx-icon>
-    <a href="#">Current</a>
-  </nav>
-</div>
-
-{% code 'html' %}
-  <nav class="hxBreadcrumb">
-    <a href="#">Home</a>
-  </nav>
-
-  <nav class="hxBreadcrumb">
-    <a href="#">Home</a>
-    <hx-icon class="delimiter" type="angle-right"></hx-icon>
-    <a href="#">Library</a>
-  </nav>
-
-  <nav class="hxBreadcrumb">
-    <a href="#">Home</a>
-    <hx-icon class="delimiter" type="angle-right"></hx-icon>
-    <a href="#">Library</a>
-    <hx-icon class="delimiter" type="angle-right"></hx-icon>
-    <a href="#">Current</a>
-  </nav>
-{% endcode %}
-{% endblock %}
