+++ docs/components/navigation/index.html
@@ -1,9 +1,5 @@
 ---
 title: Navigation
-assets:
-  - bootstrap.helix.css
-  - helix-ui.css
-  - helix-ui.js
 ---
 {% extends 'component.njk' %}
 {% block content %}
