+++ docs/components/panels/panel-demo.js
