--- docs/components/popover/index.html
@@ -1,9 +1,0 @@
----
-title: Popover
----
-{% extends 'component.njk' %}
-{% block content %}
-<!-- explanation goes here -->
-<p>TBD</p>
-
-{% endblock %}
