--- docs/components/reveal/index.html
@@ -1,61 +1,0 @@
----
-title: Reveal
-assets:
-  - helix-ui.js
----
-{% extends 'component.njk' %}
-{% block content %}
-<p>
-  A reveal is a behavior-only web component that implements almost no styling.
-</p>
-
-<section>
-  <h2 class="hxSectionTitle" id="default-reveal">Default Reveal</h2>
-  <p>Closed by default</p>
-  <div class="demo">
-    <hx-reveal>
-      <span slot="summary">Click me to show content</span>
-      <p>You can't see me.</p>
-    </hx-reveal>
-  </div>
-  {% code 'html' %}
-    <hx-reveal>
-      <span slot="summary">Click me to show content</span>
-      <p>You can't see me.</p>
-    </hx-reveal>
-  {% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="open-reveal">Open Reveal</h2>
-  <p>Add <code>open</code> attribute to set the reveal in an open state</p>
-  <div class="demo">
-    <hx-reveal open>
-      <span slot="summary">Click me to hide content!</span>
-      <p>You should see me.</p>
-    </hx-reveal>
-  </div>
-  {% code 'html' %}
-    <hx-reveal open>
-      <span slot="summary">Click me to hide content!</span>
-      <p>You should see me.</p>
-    </hx-reveal>
-  {% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="attributes">Attributes</h2>
-  <dl>
-    <dt>open</dt>
-    <dd>Opens the reveal</dd>
-  </dl>
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="properties">Properties</h2>
-  <dl>
-    <dt>open <small>(Boolean)</small></dt>
-    <dd>Determines if reveal is open.</dd>
-  </dl>
-</section>
-{% endblock %}
