--- docs/components/tabset/index.html
@@ -1,87 +1,0 @@
----
-title: Tabset
-assets:
-  - helix-ui.css
-  - helix-ui.js
----
-{% extends 'component.njk' %}
-{% block content %}
-<!-- explanation goes here -->
-
-<section>
-<h2 class="hxSectionTitle" id="tabset">Tabset</h2>
-<div class="demo">
-  <hx-tabset>
-    <div slot="tabs">
-      <a href="#" role="tab" class="hxTab">Cupcake Ipsum</a>
-      <a href="#" role="tab" class="hxTab">Biscuit Marshmallow</a>
-      <a href="#" role="tab" class="hxTab">Caramels Marzipan</a>
-    </div>
-    <hx-tabpanel id="cupcake">
-      <div class="hxRow">
-        <div class="hxCol hxSpan-6">
-          <p>
-            Cupcake ipsum dolor sit amet bonbon topping caramels. Sesame snaps
-            gummi bears liquorice cookie chupa chups fruitcake croissant
-            chocolate topping. Brownie biscuit wafer marshmallow liquorice
-            soufflé powder jelly.
-          </p>
-        </div>
-        <div class="hxCol hxSpan-6">
-          <p>
-            Sweet roll sesame snaps danish I love jelly wafer dragée soufflé
-            cake. Cookie chocolate cake gingerbread powder icing. Ice cream
-            cotton candy gummi bears oat cake sweet pastry.
-          </p>
-        </div>
-      </div>
-    </hx-tabpanel>
-    <hx-tabpanel id="biscuit">
-      <p>
-        Biscuit marshmallow chocolate cake tiramisu cupcake. Powder liquorice jelly cheesecake chocolate bar. Chocolate cake icing pudding. Tart ice cream cupcake. Cake chocolate bar carrot cake pie marshmallow sugar plum dessert. Pudding lollipop sugar plum topping sugar plum chocolate bar jelly-o pastry donut. Candy canes icing apple pie pie.
-      </p>
-    </hx-tabpanel>
-    <hx-tabpanel id="caramel">
-      <p>
-        Caramels marzipan gummi bears. Cake chocolate bar cake jelly-o tart topping. Jelly sweet jelly soufflé jelly toffee fruitcake. Toffee tart pastry liquorice bonbon sugar plum sweet. Jelly-o tiramisu cotton candy candy canes jelly beans dragée macaroon dessert toffee. Sesame snaps danish jujubes gummies marshmallow sesame snaps chocolate cake bonbon dessert. Candy canes cupcake sugar plum jelly beans bear claw icing candy canes bear claw. Marshmallow gingerbread biscuit jelly-o.
-      </p>
-    </hx-tabpanel>
-  </hx-tabset>
-</div>
-{% code 'html' %}
-  <hx-tabset>
-    <div slot="tabs">
-      <a href="#" role="tab" class="hxTab">Cupcake Ipsum</a>
-      <a href="#" role="tab" class="hxTab">Biscuit Marshmallow</a>
-      <a href="#" role="tab" class="hxTab">Caramel Marzipan</a>
-    </div>
-    <hx-tabpanel id="cupcake">
-      <div class="hxRow">
-        <div class="hxCol hxSpan-6">...</div>
-        <div class="hxCol hxSpan-6">...</div>
-      </div>
-    </hx-tabpanel>
-    <hx-tabpanel id="biscuit">...</hx-tabpanel>
-    <hx-tabpanel id="caramel">...</hx-tabpanel>
-  </hx-tabset>
-{% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="attributes">Attributes</h2>
-  <dl>
-    <dt>selected</dt>
-    <dd>Selects the tab and panel</dd>
-  </dl>
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="properties">Properties</h2>
-  <dl>
-    <dt>panels <small>(read-only) Array&lt;Element&gt;</small></dt>
-    <dt>selected</dt>
-    <dd>Determines if tab and panel is selected</dd>
-    <dt>tabs <small>(read-only) Array&lt;Element&gt;</small></dt>
-  </dl>
-</section>
-{% endblock %}
