--- docs/components/tooltip/index.html
@@ -1,8 +1,0 @@
----
-title: Tooltip
----
-{% extends 'component.njk' %}
-{% block content %}
-<!-- explanation goes here -->
-<p>TBD</p>
-{% endblock %}
