+++ docs/docs.less
@@ -68,6 +68,11 @@ main {
 @import 'demo/tooltip-docs';
 @import 'demo/typography-docs';
 
+// OVERRIDES
+.hxApp__nav a {
+  text-transform: none; // allow for <hx-icon> not <Hx-icon>
+}
+
 // SHAME
 dd {
   margin-bottom: 0.5em;
