+++ docs/elements/hx-icon/index.html
@@ -1,0 +1,97 @@
+---
+title: <hx-icon>
+---
+{% extends 'element.njk' %}
+{% block content %}
+
+<p>
+  The <code>&lt;hx-icon&gt;</code> element delivers consistent, stylable,
+  icons for use in Helix applications.
+</p>
+
+<section>
+  <h2 class="hxSectionTitle" id="syntax">Example</h2>
+  <p>
+    Set the <code>type</code> attribute to any value <a href="#icon-types">listed below</a>.
+  </p>
+  <div class="demo">
+    <p>
+      <hx-icon type="kbd-arrow-up"></hx-icon>
+      <hx-icon type="kbd-arrow-up"></hx-icon>
+      <hx-icon type="kbd-arrow-down"></hx-icon>
+      <hx-icon type="kbd-arrow-down"></hx-icon>
+      <hx-icon type="kbd-arrow-left"></hx-icon>
+      <hx-icon type="kbd-arrow-right"></hx-icon>
+      <hx-icon type="kbd-arrow-left"></hx-icon>
+      <hx-icon type="kbd-arrow-right"></hx-icon>
+      <kbd>B</kbd>
+      <kbd>A</kbd>
+      <kbd>start</kbd>
+    </p>
+  </div>
+  {% code 'html' %}
+    <hx-icon type="kbd-arrow-up"></hx-icon>
+    <hx-icon type="kbd-arrow-up"></hx-icon>
+    <hx-icon type="kbd-arrow-down"></hx-icon>
+    <hx-icon type="kbd-arrow-down"></hx-icon>
+    <hx-icon type="kbd-arrow-left"></hx-icon>
+    <hx-icon type="kbd-arrow-right"></hx-icon>
+    <hx-icon type="kbd-arrow-left"></hx-icon>
+    <hx-icon type="kbd-arrow-right"></hx-icon>
+    <kbd>B</kbd>
+    <kbd>A</kbd>
+    <kbd>start</kbd>
+  {% endcode %}
+</section>
+
+<section>
+  <h2 class="hxSectionTitle" id="styling">Styling</h2>
+  <div class="hxAlert">
+    <div class="hxAlert__icon">
+      <hx-icon type="info-circle"></hx-icon>
+    </div>
+    <div class="hxAlert__text">
+      <span class="hxAlert__status">NOTE</span>
+      Icons inherit the <code>color</code> and <code>background-color</code>
+      of their parent.
+    </div>
+  </div>
+  <br />
+
+  <p>
+    Styling Icons is similar to styling text. You can change properties like,
+    <code>font-size</code>, <code>color</code>, <code>background-color</code>,
+    <code>height</code>, <code>width</code>, etc.
+  </p>
+
+  <div class="demo">
+    <hx-icon type="times-circle"></hx-icon>
+    <hx-icon class="size-2x" type="times-circle"></hx-icon>
+    <hx-icon class="size-3x danger" type="times-circle"></hx-icon>
+  </div>
+  {% code 'html' %}
+    <hx-icon type="times-circle"></hx-icon>
+    <hx-icon class="size-2x" type="times-circle"></hx-icon>
+    <hx-icon class="size-3x danger" type="times-circle"></hx-icon>
+  {% endcode %}
+</section>
+
+<section>
+  <h2 class="hxSectionTitle" id="icon-types">Icon Types</h2>
+  <p>There are {{ data.icons.length }} icons available!</p>
+  <div class="hxRow">
+    {% for icon in data.icons %}
+      <div class="hxCol hxSpan-12-xs hxSpan-6-sm hxSpan-4-md hxSpan-3-xl">
+        <div class="demo icon-demo no-label">
+          <hx-icon type="{{icon}}"></hx-icon>
+          <p>
+            <code>{{icon}}</code>
+          </p>
+        </div>
+      </div>
+    {% else %}
+      <p>No Icons defined</p>
+    {% endfor %}
+  </div>
+</section>
+{% endblock %}
