+++ docs/elements/hx-tooltip/index.html
@@ -1,0 +1,8 @@
+---
+title: <hx-tooltip>
+---
+{% extends 'element.njk' %}
+{% block content %}
+<!-- explanation goes here -->
+<p>TBD</p>
+{% endblock %}
