+++ docs/guides/install/index.html
@@ -2,7 +2,7 @@
 title: Installing HelixUI
 cdn: https://2d2bf231b82dfe76fe36-fa12562cfe810d69bedcc36a0ac289ef.ssl.cf1.rackcdn.com
 ---
-{% extends 'component.njk' %}
+{% extends 'guide.njk' %}
 {% block content %}
 <section>
   <h2 class="hxSectionTitle" id="assets">Assets</h2>
