+++ docs/index.html
@@ -1,4 +1,4 @@
-{% extends 'fullpage.njk' %}
+{% extends 'application.njk' %}
 
 {% block content %}
 <h1>HelixUI</h1>
