+++ lib/nunjucks/toc.filter.js
@@ -20,7 +20,7 @@ function tocFilter (str, options) {
 
     // STOP, if no TOC headings found
     if (!headings.length) {
-        return '<p><em>Unavailable</em></p>';
+        return '<p><em>No anchors found</em></p>';
     }
 
     let items = [];
@@ -35,11 +35,8 @@ function tocFilter (str, options) {
 
         items.push(`
             <li class="toc__item">
-                <a
-                    class="toc__link"
-                    onclick="document.location.hash='${anchor}';">
-
-                    <span class="toc__text">${ $(this).text() }</span>
+                <a class="toc__link" href="#${anchor}">
+                    ${ $(this).text() }
                 </a>
             </li>
         `);
@@ -47,10 +44,7 @@ function tocFilter (str, options) {
 
     return `
         <ul class="toc">${items.join('')}</ul>
-        <a
-            class="toc__top-link"
-            onclick="document.location.hash='top'">
-
+        <a class="toc__top-link" href="#top">
             Back to Top
         </a>
     `;
